+++ android/src/main/java/com/bitmovin/player/reactnative/converter/JsonConverter.kt
@@ -46,6 +46,7 @@ import com.bitmovin.player.api.media.thumbnail.ThumbnailTrack
 import com.bitmovin.player.api.media.video.quality.VideoQuality
 import com.bitmovin.player.api.metadata.Metadata
 import com.bitmovin.player.api.metadata.daterange.DateRangeMetadata
+import com.bitmovin.player.api.metadata.emsg.EventMessage
 import com.bitmovin.player.api.metadata.id3.ApicFrame
 import com.bitmovin.player.api.metadata.id3.BinaryFrame
 import com.bitmovin.player.api.metadata.id3.ChapterFrame
@@ -921,6 +922,7 @@ fun Metadata.toJson(type: String): Map<String, Any> {
 fun Metadata.Entry.toJson(): Map<String, Any> {
     return when (this) {
         is DateRangeMetadata -> this.toJson()
+        is EventMessage -> this.toJson()
         is Id3Frame -> this.toJson()
         is ScteMessage -> this.toJson()
         else -> mapOf("metadataType" to "NONE")
@@ -952,6 +954,15 @@ fun DateRangeMetadata.toJson(): Map<String, Any> {
         "plannedDuration" to plannedDuration,
     ).filterNotNullValues()
 }
+
+fun EventMessage.toJson(): Map<String, Any> = mapOf(
+    "metadataType" to "EMSG",
+    "id" to id,
+    "schemeIdUri" to schemeIdUri,
+    "value" to value,
+    "duration" to durationMs,
+    "messageData" to Base64.encodeToString(messageData, Base64.NO_WRAP)
+).filterNotNullValues()
 
 private fun Id3Frame.toJson(): Map<String, Any> {
     val base = mutableMapOf<String, Any>(